Refactor MiniGraphiQL to separate files
